requirements.txt
Flask==3.1.1
mysql-connector-python
paho_mqtt==2.1.0
